refactoring connecting info desktop popup
