+++ lib/common/l10n/app_ru.arb
@@ -5,8 +5,6 @@
     "howToGetConnectingCode": "Как получить код связывания?",
     "wrongConnectingCode": "Введенный вами код неверен",
     "tryToRepeatAfter30sec": "Повторите попытку через 30 секунд",
-    "howToGetCodeTitle": "Как получить код \nсвязывания",
-    "howToGetCodeTitleWeb": "Как получить код связывания",
     "howToGetCodeText": "Код связывания генерируется службой безопасности для каждого устройства. Код одноразовый. Если вы купили новый телефон и разлогинились обратитесь в службу безопасности",
     "address": "Адрес",
     "addressForCode": "БЦ ЭРМИТАЖ ПЛАЗА, Краснопролетарская ул., 4, Москва, каб. 56",
