+++ lib/feature/presentation/ui/connecting_info/widgets/work_mode_table.dart
